+++ src/main/java/com/poo/miapi/dto/tecnico/TecnicoResponseDto.java
@@ -2,10 +2,8 @@ package com.poo.miapi.dto.tecnico;
 
 import java.util.List;
 
-import com.poo.miapi.model.historial.IncidenteTecnico;
-
 public class TecnicoResponseDto {
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
     private String email;
@@ -21,7 +19,7 @@ public class TecnicoResponseDto {
     public TecnicoResponseDto() {
     }
 
-    public TecnicoResponseDto(int id, String nombre, String apellido, String email, String rol,
+    public TecnicoResponseDto(Long id, String nombre, String apellido, String email, String rol,
             boolean cambiarPass, boolean activo, boolean bloqueado, int fallas, int marcas,
             List<IncidenteTecnicoResponseDto> incidentes) {
         this.id = id;
@@ -38,11 +36,11 @@ public class TecnicoResponseDto {
     }
 
     // Getters y setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
